requirements.txt
altair==5.3.0
apache-libcloud==3.8.0
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-mgmt-compute==31.0.0
azure-mgmt-core==1.4.0
azure-mgmt-network==25.3.0
azure-mgmt-resource==23.1.1
blinker==1.8.2
boto3==1.34.101
botocore==1.34.101
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
circuitbreaker==1.4.0
click==8.1.7
cryptography==42.0.7
gitdb==4.0.11
GitPython==3.1.43
globus-sdk==3.41.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-compute==1.19.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.64.1
idna==3.7
importlib-resources==6.4.0
isodate==0.6.1
jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msal==1.28.0
msal-extensions==1.1.0
numpy==1.24.4
oci==2.128.2
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pkgutil-resolve-name==1.3.10
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydeck==0.9.1
pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
six==1.16.0
smmap==5.0.1
streamlit==1.35.0
streamlit-autorefresh==1.0.1
tabulate==0.9.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing-extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
watchdog==4.0.1
zipp==3.18.2
